extracts routes/controllers/services that are auth related from user*
