+++ src/controllers/authController.js
@@ -1,0 +1,37 @@
+import AuthService from '../services/authService';
+
+exports.test = async (req, res, next) => {
+    try {
+        const response = await AuthService.test();
+        return res.status(response.statusCode).json(response);
+    } catch (error) {
+        return res.status(error.statusCode).json(error);
+    }
+};
+
+exports.register = async (req, res, next) => {
+    try {
+        const response = await AuthService.register(req, res, next);
+        return res.status(response.statusCode).json(response);
+    } catch (error) {
+        return res.status(error.statusCode).json(error);
+    }
+};
+
+exports.login = async (req, res, next) => {
+    try {
+        const response = await AuthService.login(req, res, next);
+        return res.status(response.statusCode).json(response);
+    } catch (error) {
+        return res.status(error.statusCode).json(error);
+    }
+};
+
+exports.validate = async (req, res, next) => {
+    try {
+        const response = await AuthService.validate();
+        return res.status(response.statusCode).json(response);
+    } catch (error) {
+        return res.status(error.statusCode).json(error);
+    }
+};
